requirements.txt
asgiref==3.11.1
Django>=4.2,<6.0
djangorestframework==3.17.1
djangorestframework-csv==3.0.2
sqlparse==0.5.5
gunicorn
